+++ CellFabric/tests/test_default_fabric.py
@@ -1,7 +1,7 @@
 from cell_fabric import DefaultCanvas, Pdk
 
 import pytest
-import fabric_default
+import fabric_CMC_NMOS_v1_default ### This is most updtaed cell fabric, therefore, replaced the fabric_default 
 import json
 import itertools
 
@@ -46,24 +46,25 @@ def test_beol_stack(setup):
 def test_fabric_default():
     unit_cap = 10
 
-    fin_u1 = 8
-    x_cells = 2
-    y_cells = 1
+    x_cells = 4	
+    y_cells = 2
+    fin_u = 12
+    #gate_u = int(sys.argv[4])
+    gate = 2
+    fin = 2*((fin_u+1)//2)
+    gateDummy = 3 ### Total Dummy gates per unit cell 2*gateDummy
+    finDummy = 4  ### Total Dummy fins per unit cell 2*finDummy
 
-    gate_u = 2
-    if fin_u1%2 != 0:
-        fin_u = fin_u1 + 1
-    else:
-        fin_u = fin_u1 
+    uc = fabric_CMC_NMOS_v1_default.UnitCell( fin_u, fin, finDummy, gate, gateDummy)
 
-    uc = fabric_default.UnitCell( gate_u, fin_u, fin_u1)
+    for (x,y) in ( (x,y) for x in range(x_cells) for y in range(y_cells)):
+        uc.unit( x, y, x_cells, y_cells, fin_u, fin, finDummy, gate, gateDummy)
 
-    for (x,y) in itertools.product( range(x_cells), range(y_cells)):
-        uc.unit( x, y, x_cells, y_cells, fin_u, gate_u)
+    uc.computeBbox()
 
-    data = uc.gen_data()
+    data = { 'bbox' : uc.bbox.toList(), 'globalRoutes' : [], 'globalRouteGrid' : [], 'terminals' : uc.terminals}
 
-    fn = "tests/__json_default_fabric_nmos"
+    fn = "tests/__json_cmc_nmos_big_v1_default"
 
     with open( fn + "_cand", "wt") as fp:
         fp.write( json.dumps( data, indent=2) + '\n')
@@ -75,7 +76,4 @@ def test_fabric_default():
         for (x,y) in zip( data['terminals'], data_golden['terminals']):
             x['netName'] = '_'
             y['netName'] = '_'
-            x['pin'] = '_'
-            y['pin'] = '_'
             assert x == y
-
